Bugfix: Modal fixed footer failed to follow keyboard
It seems like I managed to break the functionality during the previous PR cleanup.

The problem: observeEmbeddedFooter() would be triggered as the footer appeared,
then triggered again as the footer disappeared (as a direct result of it being moved).
In both cases, a reference was stored to the element, but as the element was (re)moved,
the reference was effectively lost. Thus, any subsequent attempt to set --keyboard-offset
would skip setting it on the footer.

TL;DR: duh!

Fixed by reintroducing the check that made logic run only for the footer being added.