🐛 Remove horizontal safe area in kirby-item
Ionic uses the predefined environment variables safe-area-inset-left and
safe-area-inset-right. Unfortunately this introduces extra padding in
ion-item in iOS Safari when rotated to landscape orientation.

Removing the safe area will only be a problem if kirby-item is used in
a context, where it touches the edge of the screen. This does not seem
to be the case. Apparently there's always a container with a max-width
small enough to leave space for the notch.

More information:
* https://developer.mozilla.org/en-US/docs/Web/CSS/env()
* https://ionicframework.com/docs/theming/advanced
* https://webkit.org/blog/7929/designing-websites-for-iphone-x/