🔧 Copy scss folder to root of core package
In order to be able to use scss files such as global styles in a project
via '@kirbydesign/core/scss' the folder is copied to the root of
the package. This can be handled by the stencil compiler, so as a
trade off, it will be done every time the project is built.
Therefore, the folder has also been added to .gitignore.