:bug: Make dropdown with popover work in safari
Adding tabindex='0' to the 'h3' element in the item for dropdown allows Safari to focus something after clicking an item in the list. Causing relatedTarget for the blur event to not be null. Which in turn makes it act as the remainding browsers - with the exception that relatedTarget is the 'h3' and not the 'ion-item' element.