🚧  Fix wrong spacing in virtual list causing 'bump'
Virtual scroll (#853) introduced wrong spacing to list items,
including the newly added virtual list. If this padding is
removed, the bump dissappears. It is probably because the
css-styling is added after the list has calculated the height of the
items, and therefore the list items have height added that the
virtual scroll algorithm cannot account for.