:white_check_mark: Add 1 ms to delay in 'open dropdown within...' test
When using popover the dropdown opens EXACTLY at openDelayInMs = 100ms. Setting the timeout to 100ms causes the timeout to run before the polling of the result. Sometimes the other way around. This results in a flaky test. Adding 1ms - causes the test to always succeed. In general i think whenTrue should be examined in the cases where the action happens EXACTLY at the timeout.