Added experimental "usePopover" input property to dropdown component (#1671)
* 💄 Add `popover` z-layer

* 💄 Add `zLayer`method

* ✨ Add Popover component

* 💄 Add Popover context styling

* ✨ Add KeyboardHandlerService

* ✨ Use KeyboardHandlerService

* 🤡 Add popover mock

* 🎨  Align naming with dropdown

* 🤡 Auto generate mock

* 🐛 Add missing breaks

* ✨ Add `usePopover` opt-in

* :bug: Add @dynamic decorator to popover component

The compiler will throw an error when building without due to the use of document.

* :recycle: Use elementRef to get reference to document

I think this is less cryptic than having @dynamic decorator in comment solve the problem.

* :fire: Remove unused imports

* :truck: Move 'KeyboardHandlerService' to component folder

I think the functionality of the service should be generalized more if it were to be located in helper. In its current form it only seems to be usable for the dropdown component.

* :recycle: Update imports of test to make them able to run again

A single test related to expand=block still fails tho.

* :lipstick: Restore min- & max-widths for expand=block

This has been accidentally changed and caused a test to fail.

* :recycle: Move dropdown types to separate file

* :fire: Remove 'static: false' from ViewChild decorators

It is default behaviour therefore i am removing it to reduce 'visual clutter' when skimming the file.

* :bulb: Add comment detailing why Renderer2.listen is not used

* :fire: Remove duplicate 'addClass of 'is-opening'

* :pencil2: Add 'fit' to 'contentCanRightOfTarget'

* :recycle: Provide popoverElement as argument in 'position...' functions

* :bug: Fix popover accidentally releasing scroll in modal

* :recycle: Rearrange properties & change how zIndex is set

In my opinion these changes makes the file more readable by not having them scattered all over.

* :recycle: Put zIndex logic in 'placePopoverAboveClosestModal'

* :recycle: Simplify positionHorizontally function

* :recycle: Simplify 'positionVertically'

* :recycle: Make 'targetElement' a computed property

Before it was set manually onChanges making it another piece of state that has to be maintained. I think this is cleaner as it is now just directly dependent on the target property.

* :recycle: Rename 'popover...' references to 'wrapper...'

It was actually the wrapper element being referenced and not the popover component itself. It was a bit confusing when reading the code; this way i think it is more clear what is actually being referred to.

* :bulb: Move comment to be more visible

* :bug: Make dropdown with popover work in safari

Adding tabindex='0' to the 'h3' element in the item for dropdown allows Safari to focus something after clicking an item in the list. Causing relatedTarget for the blur event to not be null. Which in turn makes it act as the remainding browsers - with the exception that relatedTarget is the 'h3' and not the 'ion-item' element.

* :label: Add types for values with implicit 'any'

* :bug: Fix blue outline bug

* :white_check_mark: Mock out popover, add default usePopover test

* :white_check_mark: Add a separate test file for dropdown with popover

See explanation in comment located at the top of the file.

* :white_check_mark: Fix "options should be visible" test

* :fire: Remove "Should have popover disabled" test

* :bug: Solve wrapper being wrongly positioned

* :white_check_mark: Update "should align to the right side..." test

* :white_check_mark: Fix 'open card to the right...' test

* :white_check_mark: Fix 'open card to the left...' test

* :white_check_mark: Fix 'should have correct item size' test

* :white_check_mark: Fix 'should render dropdown with full width' test

* Revert ":bug: Fix blue outline bug"

This reverts commit e36f8c52cb5ea9cc346354ad82b17246a1e77486.

Unfortunately the solution broke other functionality such as being able to
use esc to close. This in turn made a bunch of tests fail.

Reverting such that a better solution can be found.

* :white_check_mark: Add 1 ms to delay in 'open dropdown within...' test

When using popover the dropdown opens EXACTLY at openDelayInMs = 100ms. Setting the timeout to 100ms causes the timeout to run before the polling of the result. Sometimes the other way around. This results in a flaky test. Adding 1ms - causes the test to always succeed. In general i think whenTrue should be examined in the cases where the action happens EXACTLY at the timeout.

* :recycle: Create integration tests file, move tests there

* :recycle: Replace the use of 'unset' with 'initial'

* :fire: Remove TODO comments

* :recycle: Use 'initial' instead of 'unset' for card widths

* :recycle: Rename '--width' to '--kirby-card-width'

* :fire: Remove unused --kirby_internal_min-screen-width

* :recycle: Use named constants for min-width values

* :white_check_mark: Reword & add missing 'should' to tests

* :fire: Remove 'TODO' comment

* :white_check_mark: Rewrite test to use fakeAsync instead of timeouts

* :fire: Remove accidentaly left in 'fdescribe'

* :recycle: Restructure tests to have a flatter structure

* :white_check_mark: Restructure to have changedetectionstrategy tests work

Co-authored-by: Jakob Engelbrecht <jakob@basher.dk>
Co-authored-by: Jesper Kaltoft Vind <git@jesperkaltoft.dk>
Co-authored-by: RasmusKjeldgaard <raskjeldgaard@gmail.com>