🔧 Inintialize badge to test global styling
Global styles seem to work - we get correct font, and global variables.
Still no luck applying badge styles, probably due to the proxy comp
not having any style-urls for compodoc and/or storybook to tap into.
The styles all belong to the 'KirbyBadge' web comp, which sb does
not know.