:wrench: Bundle assets & create static folder for SB
This is a temporary and not very elegant solution that allows us to move forwards. I will be creating a separate issue for coming up with a better solution to this as i do not think it is particularly robust. But it allows us to continue the work on storybook for now.