fix(reorder-list): Issue related to "feedback"-layer not behaving with OnPush (#1899)
Feedback layer "did not appear" (or did so, too late - then didn't disappear,
at all) from "Re-orderlist" when containing Component was equipped with
OnPush-change detection.

This was a fault with MutationObserver not "triggering" when OnPush-change
detection was set. This will (again, most likely) occur due to
MutationObserver not triggering change detection, because event does not
originate from Component's element nor based on @Input-event.