Hotfix: fix Chart comonent vertical marker lines being unintentionally displayed in consumer projects using chart.js (#1969)
* :fire: Remove accidentally left in comment

* :fire: Remove @todo comment

While it has value i will move the desired solution to an issue of its own as to not clutter the source code.

* :fire: Remove chartjs.plugin-arker.d.ts

It contains the desired types for the marker plugin. I can't get typescript to register it. So i will be creating a seperate issue for this.

* :bug: Fix marker unintentionally shown in some cases

The solution is hacky but will work for now. I've linked to an issue focused on coming up with a proper solution in comments.

* :bookmark: Set versions to properly publish package to npm

We've had some trouble with the publish workflow which has made a mess of our versioning. To fix this, i'm manually setting the versions in this commit, which will cause the publish workflow to release: "5.0.15" and "0.0.22". Kinda messy, but we have to fix it.

* :fire: Remove accidentally committed declaration file