:construction: V3: Scroll viewport with embedded page "working"
Working in the sense that it seems like the items that should be rendered are actually added to the DOM, and they can be scrolled. Can be verified by inspecting down to the kirby-list-experimental level and seeing the items be swapped when scrolling far enough. But they are now shown, probably because the ion-content added by kirby page gets no height.