:bug: Handle warnings when no embedded component is found
Due to ModalConfig.component having type 'any' well... Any value is valid causing the function to occasionally return null. This is not explicit in the code or properly handled. The getEmbeddedComponentElement could handle that before as it used 'lastElementSibling' & 'firstElementSibling' causing an element to always be returned as there will always be a single element (router-outlet). This has now been made explicit and is handled by both the tests and the code itself.