Add "collapsibleTitle" field to modal config (#2174)
* :recycle: Change logic of 'getEmbeddedComponentElement'

I will have to introduce a new child element to the start of 'ion-content'. I've changed the function to search from the last element and then go to the previous sibling instead.

* :sparkles: Make it possible to toggle condensed title

* :sparkles: Allow use of list in elementToParentMap

The modal long title feature requires that the title is inserted in both the content & the ion-header element. Therefore extend with ability to use a list and allow it to be cloned to multiple new places.

* :sparkles: Let modal config control title collapse setting

This applies the setting to every page in the modal. If we need to control it on a per page basis we should allow kirby-page-title to control it.

* :recycle: Redo getEmbeddedComponentElement to avoid flakey test

The previous implementation caused tests to be flakey. This does not do so in my experience.

* :white_check_mark: Refresh test builder for each test

The testing problem stemmed from the testbuilders config.component field carrying it's value across tests. Causing config.component to sometimes be null other times to be a component.

* :bug: Handle warnings when no embedded component is found

Due to ModalConfig.component having type 'any' well... Any value is valid causing the function to occasionally return null. This is not explicit in the code or properly handled. The getEmbeddedComponentElement could handle that before as it used 'lastElementSibling' & 'firstElementSibling' causing an element to always be returned as there will always be a single element (router-outlet). This has now been made explicit and is handled by both the tests and the code itself.

* :memo: Add 'collapse title' toggle to modal configurator

* :memo: Add 'collapseTitle' to api documentation

* :memo: Make is possible to display footer as inline

* :lipstick: Remove padding & make collapsed title a h3

* :fire: Remove accidentally left in span element

* :rotating_light: Run prettier to solve tslint errors

* :rotating_light: Fix prettier errors

* :white_check_mark: Add 'no top padding' test for collapsible title

* :white_check_mark: Add tests checking if title is placed correctly

* :recycle: Pass test title as input property

* :white_check_mark: make test actually check if title is rendered

* :white_check_mark: Remove usage of ModalConfig.title

* :label: Remove usage of ModalConfig.title

* :fire: Remove unused title argument from 'openModal'

* :fire: Remove 'title' from ModalConfig interface

* :lipstick: Change font to 'P - bold;

* :fire: Remove accidentally left in fdescribe

* :recycle: Remove unnecessary instatiation of TestBuilder

* :recycle: Change 'e' to 'element'

Just writing 'e' makes it seem like some kind of error handling where that notation often is used.

* :memo: Add note regarding collapseTitle availability