Fix popovers hide() being called twice (#2195)
* 🐛  Only remove popover element if it is showing

* ♻️ Refactor hide() to ensure willHide is fired

* 🔥 Remove backdrop click listener

The event listener was not called because dropdowns onBlur event was
always fired first, closing the dropdown including the popover
component. In fact it was only called when clicking inside the dropdown
because onBlur is not fired then - and thus it is useless at this point.

Components using popover (dropdown only right now) should take care of
calling hide() at the appropriate time themselves, e.g. in onBlur.

* 🔥 Remove redundant isShowing check