🐛  Run manual CD when changing items in list
Because kirby-item is projected into the list we currently query
all the items in the list and manually set their selectable state,
because we can't bind to the input of item in the list template.

This means that we need to manually trigger change detection
when we have change the state of the item component, for it to
sync with the view correctly, and become clickable.