💄 Change how backdrop-opacity is applied to modal
This change allows us to remove all the custom animation logic
that was copied from Ionics code base, and enhanced with some
logic for reducing the opacity of underlying modals. This was no longer
working with Ionic 6, and had to be refactored or removed.

The cost of this change is that modal backdrop opacity now stacks,
so the background gets darker and darker with nested modals.