♻️  Refactor ActionSheet to preserve backdrop click
We are scaling and moving the modal in unexpected ways to position
the action sheet. Previously we were able to hide the modal shadow
element that has the wrong size and position via CSS but it is now
encapsulated and not exposed via a part in Ionic 6. Therefore it is
removed instead.