:bug: Fix slotted items being unclickable after update
Fixes #2153 - the issue was caused by the event listener being set up once using the renderer on afterContentChecked. If the slotted items changed an if sentence did so the new items did not have their click listener set up. Now we do it everytime the kirbyItemsSlotted attribute is updated. We also make sure to unlisten old items when it changes & onDestroy.