Upgrade Sass and update all tilde imports
Tilde imports have been deprecated in Sass loader for a long time and
will throw warnings when we upgrade to Angular 14.
Removing the tilde charater `~` from the path in all @use statements
fixes this.

By doing this in a separate PR when can reduce clutter in the Angular 14
upgrade PR.